+++ pkg/server/api/console/v1/cookie_banner_resolvers.go
@@ -727,13 +727,14 @@ func (r *mutationResolver) UpdateCookieBanner(ctx context.Context, input types.U
 		ctx,
 		scope,
 		cookiebanner.UpdateCookieBannerRequest{
-			CookieBannerID:    input.CookieBannerID,
-			Name:              input.Name,
-			PrivacyPolicyURL:  input.PrivacyPolicyURL,
-			CookiePolicyURL:   input.CookiePolicyURL,
-			ConsentExpiryDays: input.ConsentExpiryDays,
-			DefaultLanguage:   input.DefaultLanguage,
-			Capabilities:      capabilities,
+			CookieBannerID:       input.CookieBannerID,
+			Name:                 input.Name,
+			PrivacyPolicyURL:     input.PrivacyPolicyURL,
+			CookiePolicyURL:      input.CookiePolicyURL,
+			ConsentExpiryDays:    input.ConsentExpiryDays,
+			DefaultLanguage:      input.DefaultLanguage,
+			PublisherCountryCode: input.PublisherCountryCode,
+			Capabilities:         capabilities,
 		},
 	)
 	if err != nil {
@@ -958,6 +959,11 @@ func (r *mutationResolver) UpdateCookieCategory(ctx context.Context, input types
 		gcmConsentTypes = &input.GcmConsentTypes
 	}
 
+	var tcfPurposeIDs *[]int
+	if input.TcfPurposeIds != nil {
+		tcfPurposeIDs = &input.TcfPurposeIds
+	}
+
 	category, err := r.cookieBanner.UpdateCookieCategory(
 		ctx,
 		scope,
@@ -967,6 +973,7 @@ func (r *mutationResolver) UpdateCookieCategory(ctx context.Context, input types
 			Slug:             input.Slug,
 			Description:      input.Description,
 			GCMConsentTypes:  gcmConsentTypes,
+			TCFPurposeIDs:    tcfPurposeIDs,
 			PostHogConsent:   input.PosthogConsent,
 		},
 	)